fix[devtools/inspectElement]: dont pause initial inspectElement call when user switched tabs
