+++ packages/react-devtools-shared/src/devtools/views/Components/InspectedElementSuspenseToggle.js
@@ -11,10 +11,10 @@ import * as React from 'react';
 import {OptionsContext} from '../context';
 import EditableValue from './EditableValue';
 import Store from '../../store';
-import {ElementTypeSuspense} from 'react-devtools-shared/src/types';
+import {ElementTypeSuspense} from 'react-devtools-shared/src/frontend/types';
 import styles from './InspectedElementSharedStyles.css';
 
-import type {InspectedElement} from './types';
+import type {InspectedElement} from 'react-devtools-shared/src/frontend/types';
 import type {FrontendBridge} from 'react-devtools-shared/src/bridge';
 
 type Props = {
