+++ packages/react-devtools-shared/src/devtools/views/Components/NewKeyValue.js
@@ -15,7 +15,7 @@ import EditableValue from './EditableValue';
 import {parseHookPathForEdit} from './utils';
 import styles from './NewKeyValue.css';
 
-import type {InspectedElement} from './types';
+import type {InspectedElement} from 'react-devtools-shared/src/frontend/types';
 import type {FrontendBridge} from 'react-devtools-shared/src/bridge';
 
 type Props = {
