+++ packages/react-devtools-shared/src/devtools/views/Profiler/RankedChartBuilder.js
@@ -10,7 +10,7 @@
 import {
   ElementTypeForwardRef,
   ElementTypeMemo,
-} from 'react-devtools-shared/src/types';
+} from 'react-devtools-shared/src/frontend/types';
 import {formatDuration} from './utils';
 import ProfilerStore from 'react-devtools-shared/src/devtools/ProfilerStore';
