+++ packages/react-devtools-shared/src/hydration.js
@@ -14,9 +14,12 @@ import {
   getInObject,
   formatDataForPreview,
   setInObject,
-} from './utils';
+} from 'react-devtools-shared/src/utils';
 
-import type {DehydratedData} from './devtools/views/Components/types';
+import type {
+  DehydratedData,
+  InspectedElementPath,
+} from 'react-devtools-shared/src/frontend/types';
 
 export const meta = {
   inspectable: (Symbol('inspectable'): symbol),
@@ -123,7 +126,7 @@ export function dehydrate(
   unserializable: Array<Array<string | number>>,
   path: Array<string | number>,
   isPathAllowed: (path: Array<string | number>) => boolean,
-  level?: number = 0,
+  level: number = 0,
 ): $PropertyType<DehydratedData, 'data'> {
   const type = getDataType(data);
 
@@ -365,7 +368,7 @@ export function dehydrate(
 export function fillInPath(
   object: Object,
   data: DehydratedData,
-  path: Array<string | number>,
+  path: InspectedElementPath,
   value: any,
 ) {
   const target = getInObject(object, path);
